+++ src/modula.ts
@@ -1,4 +1,6 @@
 import Component from './Components/Component';
+import PageComponent from './Components/PageComponent';
+import TemplateComponent from './Components/TemplateComponent';
 import html from './Core/Template';
 import IComponentDefinition from './Interface/IComponentDefinition';
 import IModulaOptions from './Interface/IModulaOptions';
@@ -7,6 +9,8 @@ import Modula from './Core/Modula';
 
 export {
     Component,
+    PageComponent,
+    TemplateComponent,
     html,
     IComponentDefinition,
     IModulaOptions,